Move converting frozen patcher data to dict into freezing_utils
