+++ maxdiff/freezing_utils.py
@@ -1,3 +1,4 @@
+import json
 import datetime
 from typing import Optional
 
@@ -63,3 +64,45 @@ def get_hfs_date(data: bytes) -> datetime.datetime:
     return datetime.datetime.fromtimestamp(
         seconds_in_hfs_plus - seconds_offset_from_unix, datetime.UTC
     )
+
+
+def get_patcher_dict(entry: device_entry_with_data):
+    """Returns the dict that represents the given patcher data.
+    Prints errors if parsing fails"""
+
+    if not "data" in entry:
+        return {}
+
+    patch_data = entry["data"]
+    if not isinstance(patch_data, bytes):
+        return {}
+
+    if not "file_name" in entry:
+        return {}
+
+    name = entry["file_name"]
+    if not isinstance(name, str):
+        return {}
+
+    device_data_text = ""
+    try:
+        if patch_data[len(patch_data) - 1] == 0:
+            device_data_text = patch_data[: len(patch_data) - 1].decode("utf-8")
+        else:
+            device_data_text = patch_data.decode("utf-8")
+    except Exception as e:
+        print(f"Error getting patch data as text for entry {name}: {e}")
+        return {}
+
+    try:
+        patcher_dict = json.loads(device_data_text)
+    except ValueError as e:
+        print(f"Error parsing device patch data as json for entry {name}: {e}")
+        return {}
+
+    try:
+        patcher = patcher_dict["patcher"]
+        return patcher
+    except:
+        print(f"Content of entry {name} does not seem to be a patcher")
+        return {}
